Restructured development pages and added plans for 2019 and 2020
Started to restructure the website to clearly separate the subsections
so that users can more easily find and read about only the parts that
interest them. In this first phase of the restructuring moved all the
of the developer information to a subsite.

In addition, add some of the planning information from 2019, along
with the progress made, and create about half of the planning
information for 2020.

Both the restructuring and the planning will continue for the next
couple weeks.